feat(html): support hooks
