+++ crates/rspack_plugin_html/src/lib.rs
@@ -1,11 +1,15 @@
 #![feature(box_patterns)]
 
+pub mod asset;
 pub mod config;
+pub mod injector;
 pub mod parser;
 pub mod sri;
-pub mod visitors;
+pub mod tag;
+pub mod template;
 
 mod drive;
 mod plugin;
+
 pub use drive::*;
 pub use plugin::*;
