--- crates/rspack_plugin_html/src/visitors/mod.rs
@@ -1,3 +1,0 @@
-pub mod asset;
-pub mod tag;
-pub mod utils;
